Move dequant chunking regressions into test_kernels.py
Group the two workspace-bounding tests with the other kernel-level CUDA
tests instead of a dedicated module, and widen the reference-comparison
atol to 5e-2: the chunked and full-size fp16 GEMMs may pick different
CUDA algorithms, and an RTX 5060 Ti run showed a 0.039 absolute
difference that the previous 1e-2 atol rejected.
